requirements.txt
streamlit
numpy
tensorflow
scikit-image
opencv-python-headless
Pillow
opencv-python
pandas
matplotlib
